Refactor: move the responsibility of throw an exception when removing or updating no existing documents to the Service

+++ Papyrus.Business/Documents/DocumentService.cs
@@ -30,11 +30,15 @@
         {
             if (string.IsNullOrWhiteSpace(document.Id))
                 throw new DocumentIdMustBeDefinedException();
+            if (await GetDocumentById(document.Id) == null)
+                throw new DocumentNotFoundException();
             await repository.Update(document);
         }
 
-        public async Task Remove(string documentId)
+        public virtual async Task Remove(string documentId)
         {
+            if (await GetDocumentById(documentId) == null)
+                throw new DocumentNotFoundException();
             await repository.Delete(documentId);
         }
